feat: 기존 MessageInput과 연동하여 메시지 작성 기능 구현
